+++ scripts/components/responsive-slider/responsive-slider.js
@@ -109,7 +109,7 @@ export const ResponsiveSlider = (props) => {
 
 				const nearestValid = getNearest();
 
-				const parsedValue = parseInt(inheritCheck(currentValue) ? nearestValid : currentValue);
+				const parsedValue = parseFloat(inheritCheck(currentValue) ? nearestValid : currentValue);
 
 				return (
 					<div className='es-h-spaced' key={index}>
